Reading of ADS state

+++ src/test/scala/com/vroste/adsclient/MyStruct.scala
@@ -1,9 +1,8 @@
 package com.vroste.adsclient
 
 import scodec.Codec
-import AdsCodecs._
 
-case class MyStruct(myInt: Int, myBool: Boolean)
+case class MyStruct(myInt: Short, myBool: Boolean)
 
 object MyStruct {
   implicit val codec: Codec[MyStruct] = (int :: bool).as
